changes to card endpoint

+++ app/Http/Controllers/CardController.php
@@ -5,19 +5,11 @@ namespace App\Http\Controllers;
 use App\Http\Resources\CardResource;
 use App\Models\Boards;
 use App\Models\Card;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 
 class CardController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
-    public function index(Boards $board)
-    {
-        $card = $board->cards()->latest()->get();
-
-        return CardResource::collection($card);
-    }
 
     /**
      * Store a newly created resource in storage.
@@ -35,21 +27,14 @@ class CardController extends Controller
         return new CardResource($card);
     }
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(Boards $board,Card $card)
-    {
-        $card = $board->cards()->findOrFail($card->id);
-        return new CardResource($card);
-    }
 
     /**
      * Update the specified resource in storage.
      */
     public function update(Request $request, Boards $board, Card $card)
     {
-        $card = $board->cards()->findOrFail($card->id);
+        Gate::authorize('update', $card);
+
         $card->update(
             $request->validate([
                 'title' => 'required|string|max:50',
@@ -65,8 +50,10 @@ class CardController extends Controller
      */
     public function destroy(Boards $board, Card $card)
     {
-        $card = $board->cards()->findOrFail($card->id);
+        Gate::authorize('delete', $card);
+
         $card->delete();
+        
         return response(status: 204);
     }
 }
